Change folder structure and add readme
Based on stellar/go layout
