+++ packages/demo/bridge/config/key/errors.go
@@ -1,0 +1,9 @@
+package key
+
+import "fmt"
+
+// Errors
+var (
+	ErrInvalidEthereumPrivateKey = fmt.Errorf("Ethereum: Invalid private key format")
+	ErrInvalidStellarSecret      = fmt.Errorf("Stellar: Invalid secret format")
+)
