+++ packages/demo/bridge/config/key/internal.go
@@ -1,0 +1,6 @@
+package key
+
+const (
+	promptEthPrivateKey = "Enter Ethereum private key: "
+	promptStellarSecret = "Enter Stellar secret: "
+)
